layout/base/tests/test_bug332655-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=332655
-->
<head>
  <title>Test for Bug 332655</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="test()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=332655">Mozilla Bug 332655</a>
<p id="display"></p>
<div id="content">
<input type="text" id="testInput"
       style="-moz-appearance: none"> <!-- bug 1234659 workaround -->
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 332655 **/

SimpleTest.waitForExplicitFinish();

function test() {
  var textInput = $("testInput");
  var s1, s2, s3, equal, str1, str2;

  textInput.focus();
  sendString("\u05d0");
  sendString("ab ");
  sendString("\u05d1\u05d2");
  s1 = snapshotWindow(window);

  // 4 LEFT to get to the beginning of the line: HOME doesn't work on OS X
  synthesizeKey("KEY_ArrowLeft", {repeat: 4});
  synthesizeKey("KEY_Backspace");
  sendString("\u05d0");
  s2 = snapshotWindow(window);

  [equal, str1, str2] = compareSnapshots(s1, s2, true);
  ok(equal, "deleting and inserting RTL char at beginning of line shouldn't change: expected " +
     str1 + " but got " + str2);

  textInput.select();
  sendString("ab ");
  sendString("\u05d1\u05d2");
  // 4 LEFT to get to the beginning of the line: HOME doesn't work on OS X
  synthesizeKey("KEY_ArrowLeft", {repeat: 4});
  sendString("\u05d0");

  s3 = snapshotWindow(window);

  [equal, str1, str2] = compareSnapshots(s1, s3, true);
  ok(equal, "the order entering Bidi text shouldn't change rendering: expected " +
     str1 + " but got " + str2);

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
